renamed and moved openEHR related classes, fixed resource leak bug
