+++ dsf-bpe/dsf-bpe-process-base/src/main/java/org/highmed/dsf/openehr/client/OpenEhrWebserviceClientProvider.java
@@ -1,10 +1,10 @@
 package org.highmed.dsf.openehr.client;
 
-import org.highmed.openehr.client.OpenehrWebserviceClient;
+import org.highmed.openehr.client.OpenEhrWebserviceClient;
 
 public interface OpenEhrWebserviceClientProvider
 {
 	String getBaseUrl();
 
-	OpenehrWebserviceClient getWebserviceClient();
+	OpenEhrWebserviceClient getWebserviceClient();
 }
