+++ dsf-openehr/dsf-openehr-model/src/main/java/org/highmed/openehr/json/RowElementDeserializer.java
@@ -1,4 +1,4 @@
-package org.highmed.openehr.deserializer;
+package org.highmed.openehr.json;
 
 import java.io.IOException;
 import java.time.ZonedDateTime;
