+++ dsf-openehr/dsf-openehr-webservice-client/src/main/java/org/highmed/openehr/client/OpenEhrWebserviceClient.java
@@ -4,7 +4,7 @@ import javax.ws.rs.core.MultivaluedMap;
 
 import org.highmed.openehr.model.structure.ResultSet;
 
-public interface OpenehrWebserviceClient
+public interface OpenEhrWebserviceClient
 {
 	ResultSet query(String query, MultivaluedMap<String, Object> headers);
 }
