+++ dsf-openehr/dsf-openehr-webservice-client/src/test/java/org/highmed/openehr/client/TestOpenEhrJerseyClient.java
@@ -1,16 +1,16 @@
 package org.highmed.openehr.client;
 
-import org.highmed.openehr.OpenEhrObjectMapperFactory;
+import org.highmed.openehr.json.OpenEhrObjectMapperFactory;
 import org.highmed.openehr.model.datatypes.JsonNodeRowElement;
 import org.highmed.openehr.model.structure.ResultSet;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
-public class TestOpenehrJerseyClient
+public class TestOpenEhrJerseyClient
 {
 	public static void main(String... args)
 	{
-		OpenehrWebserviceClient client = new OpenehrWebserviceClientJersey("http://localhost:8003/rest/openehr/v1",
+		OpenEhrWebserviceClient client = new OpenEhrWebserviceClientJersey("http://localhost:8003/rest/openehr/v1",
 				"username", "password", 2500, 5000, objectMapper());
 
 		String query = "SELECT e FROM EHR e";
